backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pandas>=2.2.0
sqlalchemy>=2.0.25
aiosqlite>=0.19.0
httpx>=0.26.0
python-dotenv>=1.0.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
apscheduler>=3.10.4
fredapi>=0.5.1
pyyaml>=6.0
greenlet>=3.0.0
anthropic>=0.18.0
openai>=1.12.0
redis>=5.0.0
aioredis>=2.0.0
